todo_back/requirements.txt
django==3.1.1
django-rest-framework==3.11.1